telugu-corpuseum/TeluguCorpuseum/requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
importlib_metadata==7.2.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==2.1.1
numpy==1.26.4
packaging==23.2
pandas==2.3.1
pillow==10.4.0
protobuf==4.25.8
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==13.9.4
rpds-py==0.27.0
six==1.17.0
smmap==5.0.2
streamlit==1.30.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
zipp==3.23.0
